+++ tests/test_schema.py
@@ -4,12 +4,16 @@ from typing import Any, cast
 
 import pystac
 import pytest
+from jsonschema.exceptions import ValidationError
 from pystac.validation.stac_validator import STACValidator
 
 from stac_model.base import JSON
 from stac_model.schema import SCHEMA_URI
 
 from conftest import get_all_stac_item_examples
+
+# ignore typing errors introduced by generic JSON manipulation errors
+# mypy: disable_error_code="arg-type,call-overload,index,union-attr"
 
 
 @pytest.mark.parametrize(
@@ -32,7 +36,7 @@ def test_mlm_schema(
     ["item_raster_bands.json"],
     indirect=True,
 )
-def test_mlm_no_undefined_prefixed_field(
+def test_mlm_no_undefined_prefixed_field_item_properties(
     mlm_validator: STACValidator,
     mlm_example: dict[str, JSON],
 ) -> None:
@@ -40,11 +44,82 @@ def test_mlm_no_undefined_prefixed_field(
     mlm_item = pystac.Item.from_dict(mlm_data)
     pystac.validation.validate(mlm_item, validator=mlm_validator)  # ensure original is valid
 
-    mlm_data["properties"]["mlm:unknown"] = "random"  # type: ignore
+    # undefined property anywhere in the schema
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_data["properties"]["mlm:unknown"] = "random"
     with pytest.raises(pystac.errors.STACValidationError) as exc:
         mlm_item = pystac.Item.from_dict(mlm_data)
         pystac.validation.validate(mlm_item, validator=mlm_validator)
-    assert all(field in str(exc.value.source) for field in ["mlm:unknown", "^(?!mlm:)"])
+    assert all(
+        info in str(exc.value.source)
+        for info in ["mlm:unknown", "^(?!mlm:)"]
+    )
+
+    # defined property only allowed at the Asset level
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_data["properties"]["mlm:artifact_type"] = "torch.save"
+    with pytest.raises(pystac.errors.STACValidationError) as exc:
+        mlm_item = pystac.Item.from_dict(mlm_data)
+        pystac.validation.validate(mlm_item, validator=mlm_validator)
+    errors = cast(list[ValidationError], exc.value.source)
+    assert "mlm:artifact_type" in str(errors[0].validator_value)
+    assert errors[0].schema["description"] == "Fields that are disallowed under the Item properties."
+
+
+@pytest.mark.parametrize(
+    "mlm_example",
+    ["item_raster_bands.json"],
+    indirect=True,
+)
+@pytest.mark.parametrize(
+    ["test_field", "test_value"],
+    [
+        ("mlm:unknown", "random"),
+        ("mlm:name", "test-model"),
+        ("mlm:input", []),
+        ("mlm:output", []),
+        ("mlm:hyperparameters", {"test": {}}),
+    ]
+)
+def test_mlm_no_undefined_prefixed_field_asset_properties(
+    mlm_validator: STACValidator,
+    mlm_example: dict[str, JSON],
+    test_field: str,
+    test_value: Any,
+) -> None:
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    pystac.validation.validate(mlm_item, validator=mlm_validator)  # ensure original is valid
+    assert mlm_data["assets"]["weights"]
+
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_data["assets"]["weights"][test_field] = test_value
+    with pytest.raises(pystac.errors.STACValidationError) as exc:
+        mlm_item = pystac.Item.from_dict(mlm_data)
+        pystac.validation.validate(mlm_item, validator=mlm_validator)
+    assert len(exc.value.source) == 1
+    errors = cast(list[ValidationError], exc.value.source)
+    assert test_field in errors[0].instance
+    assert errors[0].schema["description"] in [
+        "All possible MLM fields regardless of the level they apply (Collection, Item, Asset, Link).",
+        "Fields that are disallowed under the Asset properties."
+    ]
+
+
+@pytest.mark.parametrize(
+    "mlm_example",
+    ["item_raster_bands.json"],
+    indirect=True,
+)
+def test_mlm_allowed_field_asset_properties_override(
+    mlm_validator: STACValidator,
+    mlm_example: dict[str, JSON],
+) -> None:
+    # defined property allowed both at the Item at the Asset level
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_data["assets"]["weights"]["mlm:accelerator"] = "cuda"
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    pystac.validation.validate(mlm_item, validator=mlm_validator)
 
 
 @pytest.mark.parametrize(
@@ -60,7 +135,7 @@ def test_mlm_missing_bands_invalid_if_mlm_input_lists_bands(
     pystac.validation.validate(mlm_item, validator=mlm_validator)  # ensure original is valid
 
     mlm_bands_bad_data = copy.deepcopy(mlm_example)
-    mlm_bands_bad_data["assets"]["weights"].pop("raster:bands")  # type: ignore  # no 'None' to raise in case modified
+    mlm_bands_bad_data["assets"]["weights"].pop("raster:bands")  # no 'None' to raise in case missing
     with pytest.raises(pystac.errors.STACValidationError):
         mlm_bands_bad_item = pystac.Item.from_dict(mlm_bands_bad_data)
         pystac.validation.validate(mlm_bands_bad_item, validator=mlm_validator)
@@ -79,7 +154,7 @@ def test_mlm_eo_bands_invalid_only_in_item_properties(
     pystac.validation.validate(mlm_item, validator=mlm_validator)  # ensure original is valid
 
     mlm_eo_bands_bad_data = copy.deepcopy(mlm_example)
-    mlm_eo_bands_bad_data["assets"]["weights"].pop("eo:bands")  # type: ignore  # no 'None' to raise in case modified
+    mlm_eo_bands_bad_data["assets"]["weights"].pop("eo:bands")  # no 'None' to raise in case missing
     with pytest.raises(pystac.errors.STACValidationError):
         mlm_eo_bands_bad_item = pystac.Item.from_dict(mlm_eo_bands_bad_data)
         pystac.validation.validate(mlm_eo_bands_bad_item, validator=mlm_validator)
@@ -95,12 +170,12 @@ def test_mlm_no_input_allowed_but_explicit_empty_array_required(
     mlm_example: dict[str, JSON],
 ) -> None:
     mlm_data = copy.deepcopy(mlm_example)
-    mlm_data["properties"]["mlm:input"] = []  # type: ignore
+    mlm_data["properties"]["mlm:input"] = []
     mlm_item = pystac.Item.from_dict(mlm_data)
     pystac.validation.validate(mlm_item, validator=mlm_validator)
 
     with pytest.raises(pystac.errors.STACValidationError):
-        mlm_data["properties"].pop("mlm:input")  # type: ignore  # no 'None' to raise in case modified
+        mlm_data["properties"].pop("mlm:input")  # no 'None' to raise in case missing
         mlm_item = pystac.Item.from_dict(mlm_data)
         pystac.validation.validate(mlm_item, validator=mlm_validator)
 
@@ -163,13 +238,13 @@ def test_mlm_other_non_mlm_assets_allowed(
     mlm_item = pystac.Item.from_dict(mlm_data)
     pystac.validation.validate(mlm_item, validator=mlm_validator)  # self-check valid beforehand
 
-    mlm_data["assets"]["sample"] = {  # type: ignore
+    mlm_data["assets"]["sample"] = {
         "type": "image/jpeg",
         "href": "https://example.com/sample/output.jpg",
         "roles": ["preview"],
         "title": "Model Output Predictions Sample",
     }
-    mlm_data["assets"]["model-cart"] = {  # type: ignore
+    mlm_data["assets"]["model-cart"] = {
         "type": "text/markdown",
         "href": "https://example.com/sample/model.md",
         "roles": ["metadata"],
@@ -184,7 +259,53 @@ def test_mlm_other_non_mlm_assets_allowed(
     ["item_basic.json"],
     indirect=True,
 )
+@pytest.mark.parametrize(
+    ["model_asset_extras", "is_valid"],
+    [
+        ({"roles": ["checkpoint"]}, False),
+        ({"roles": ["checkpoint", "mlm:model"]}, False),
+        ({"roles": ["checkpoint"], "mlm:artifact_type": "test"}, False),
+        ({"roles": ["checkpoint", "mlm:model"], "mlm:artifact_type": "test"}, True),
+    ]
+)
 def test_mlm_at_least_one_asset_model(
+    mlm_validator: STACValidator,
+    mlm_example: dict[str, JSON],
+    model_asset_extras: dict[str, Any],
+    is_valid: bool,
+) -> None:
+    mlm_data = copy.deepcopy(mlm_example)
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    pystac.validation.validate(mlm_item, validator=mlm_validator)  # self-check valid beforehand
+
+    mlm_model = {
+        "type": "application/octet-stream; application=pytorch",
+        "href": "https://example.com/sample/checkpoint.pt",
+        "title": "Model Weights Checkpoint",
+    }
+    mlm_model.update(model_asset_extras)
+    mlm_data["assets"] = {
+        "model": mlm_model  # type: ignore
+    }
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    if is_valid:
+        pystac.validation.validate(mlm_item, validator=mlm_validator)
+    else:
+        with pytest.raises(pystac.errors.STACValidationError) as exc:
+            pystac.validation.validate(mlm_item, validator=mlm_validator)
+        errors = cast(list[ValidationError], exc.value.source)
+        assert errors[0].schema["$comment"] in [
+            "At least one Asset must provide the model definition indicated by the 'mlm:model' role.",
+            "Used to check the artifact type property that is required by a Model Asset annotated by 'mlm:model' role."
+        ]
+
+
+@pytest.mark.parametrize(
+    "mlm_example",
+    ["item_basic.json"],
+    indirect=True,
+)
+def test_mlm_asset_artifact_type_checked(
     mlm_validator: STACValidator,
     mlm_example: dict[str, JSON],
 ) -> None:
@@ -192,17 +313,17 @@ def test_mlm_at_least_one_asset_model(
     mlm_item = pystac.Item.from_dict(mlm_data)
     pystac.validation.validate(mlm_item, validator=mlm_validator)  # self-check valid beforehand
 
-    mlm_data["assets"] = {  # needs at least 1 asset with role 'mlm:model'
-        "model": {
-            "type": "application/octet-stream; application=pytorch",
-            "href": "https://example.com/sample/checkpoint.pt",
-            "roles": ["checkpoint"],
-            "title": "Model Weights Checkpoint",
-        }
-    }
-    with pytest.raises(pystac.errors.STACValidationError):
-        mlm_item = pystac.Item.from_dict(mlm_data)
+    mlm_data["assets"]["model"]["mlm:artifact_type"] = 1234  # type: ignore
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    with pytest.raises(pystac.errors.STACValidationError) as exc:
         pystac.validation.validate(mlm_item, validator=mlm_validator)
+    assert "1234 is not of type 'string'" in str(exc.value.source)
+
+    mlm_data["assets"]["model"]["mlm:artifact_type"] = ""  # type: ignore
+    mlm_item = pystac.Item.from_dict(mlm_data)
+    with pytest.raises(pystac.errors.STACValidationError) as exc:
+        pystac.validation.validate(mlm_item, validator=mlm_validator)
+    assert "should be non-empty" in str(exc.value.source)
 
 
 def test_model_metadata_to_dict(eurosat_resnet):